temp

+++ external/Box2D/Collision/Shapes/b2ChainShape.h
@@ -80,7 +80,7 @@ public:
 
 	/// Chains have zero mass.
 	/// @see b2Shape::ComputeMass
-	void ComputeMass(b2MassData* massData, float32 density) const;
+	void ComputeMass(b2MassData* massData, qreal density) const;
 
 	/// The vertices. Owned by this class.
 	b2Vec2* m_vertices;
