+++ external/Box2D/Collision/Shapes/b2EdgeShape.h
@@ -49,7 +49,7 @@ public:
 	void ComputeAABB(b2AABB* aabb, const b2Transform& transform, int32 childIndex) const;
 
 	/// @see b2Shape::ComputeMass
-	void ComputeMass(b2MassData* massData, float32 density) const;
+	void ComputeMass(b2MassData* massData, qreal density) const;
 	
 	/// These are the edge vertices
 	b2Vec2 m_vertex1, m_vertex2;
