+++ external/Box2D/Dynamics/Joints/b2MotorJoint.h
@@ -41,16 +41,16 @@ struct b2MotorJointDef : public b2JointDef
 	b2Vec2 linearOffset;
 
 	/// The bodyB angle minus bodyA angle in radians.
-	float32 angularOffset;
+	qreal angularOffset;
 	
 	/// The maximum motor force in N.
-	float32 maxForce;
+	qreal maxForce;
 
 	/// The maximum motor torque in N-m.
-	float32 maxTorque;
+	qreal maxTorque;
 
 	/// Position correction factor in the range [0,1].
-	float32 correctionFactor;
+	qreal correctionFactor;
 };
 
 /// A motor joint is used to control the relative motion
@@ -62,34 +62,34 @@ public:
 	b2Vec2 GetAnchorA() const;
 	b2Vec2 GetAnchorB() const;
 
-	b2Vec2 GetReactionForce(float32 inv_dt) const;
-	float32 GetReactionTorque(float32 inv_dt) const;
+	b2Vec2 GetReactionForce(qreal inv_dt) const;
+	qreal GetReactionTorque(float32 inv_dt) const;
 
 	/// Set/get the target linear offset, in frame A, in meters.
 	void SetLinearOffset(const b2Vec2& linearOffset);
 	const b2Vec2& GetLinearOffset() const;
 
 	/// Set/get the target angular offset, in radians.
-	void SetAngularOffset(float32 angularOffset);
-	float32 GetAngularOffset() const;
+	void SetAngularOffset(qreal angularOffset);
+	qreal GetAngularOffset() const;
 
 	/// Set the maximum friction force in N.
-	void SetMaxForce(float32 force);
+	void SetMaxForce(qreal force);
 
 	/// Get the maximum friction force in N.
-	float32 GetMaxForce() const;
+	qreal GetMaxForce() const;
 
 	/// Set the maximum friction torque in N*m.
-	void SetMaxTorque(float32 torque);
+	void SetMaxTorque(qreal torque);
 
 	/// Get the maximum friction torque in N*m.
-	float32 GetMaxTorque() const;
+	qreal GetMaxTorque() const;
 
 	/// Set the position correction factor in the range [0,1].
-	void SetCorrectionFactor(float32 factor);
+	void SetCorrectionFactor(qreal factor);
 
 	/// Get the position correction factor in the range [0,1].
-	float32 GetCorrectionFactor() const;
+	qreal GetCorrectionFactor() const;
 
 	/// Dump to b2Log
 	void Dump();
@@ -106,12 +106,12 @@ protected:
 
 	// Solver shared
 	b2Vec2 m_linearOffset;
-	float32 m_angularOffset;
+	qreal m_angularOffset;
 	b2Vec2 m_linearImpulse;
-	float32 m_angularImpulse;
-	float32 m_maxForce;
-	float32 m_maxTorque;
-	float32 m_correctionFactor;
+	qreal m_angularImpulse;
+	qreal m_maxForce;
+	qreal m_maxTorque;
+	qreal m_correctionFactor;
 
 	// Solver temp
 	int32 m_indexA;
@@ -121,13 +121,13 @@ protected:
 	b2Vec2 m_localCenterA;
 	b2Vec2 m_localCenterB;
 	b2Vec2 m_linearError;
-	float32 m_angularError;
-	float32 m_invMassA;
-	float32 m_invMassB;
-	float32 m_invIA;
-	float32 m_invIB;
+	qreal m_angularError;
+	qreal m_invMassA;
+	qreal m_invMassB;
+	qreal m_invIA;
+	qreal m_invIB;
 	b2Mat22 m_linearMass;
-	float32 m_angularMass;
+	qreal m_angularMass;
 };
 
 #endif
